Update Footer, Implement theme selector, and develop team page
Updated the footer by changing the font and adding the icons to the down bottom instead of the follow us section. developed the theme selector, it gives the user the ability to change theme between light and dark. develop the team page but will be fully developed in the next commit